\name{gtbl_pph17p1a}
\alias{gtbl_pph17p1a}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate Bracket based on Income Tax Law article 17 paragraph 1 (a).
}
\description{
Generate dataset containing the minimum and maximum on each part
and tarifs.
}
\usage{
gtbl_pph17p1a(filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{CSV File}
}
\note{
CSV File format: No (Id: a numeric vector), Min (Minimum Taxable
Income: a numeric vector), Max (Maximum Taxable Income: a numeric
vector) and Tarif (numeric vector)
}
