\name{pph2p3a}
\alias{pph2p3a}
\title{
Individual Tax Resident Function
}
\description{
Subject of Individual Tax Resident based on Income Tax Article 2 paragraph 3 (a).
}
\usage{
pph2p3a(resident_date, target)
}
\arguments{
  \item{resident_date}{When the indiviuals came to Indonesia.}
  \item{target}{Target date, basically it's end of year or now on sysdate.}
}
\examples{
pph2p3a("2017-02-14")
}
